requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
requests>=2.28.0
slack-bolt>=1.18.0
twilio>=8.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
openai>=1.0.0
PyPDF2>=3.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pydantic>=2.0.0
numpy>=1.21.0